unsigned

+++ ft_hexa.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-void	ft_hexa(size_t i, char *base, int *counter)
+void	ft_hexa(unsigned i, char *base, int *counter)
 {
 	if(i >= 16)
 		ft_hexa(i / 16, base, counter);
